infrastructure/requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
boto3>=1.35.0
